refactor: add FindPokemonByTypeService on ListPokemonController
