+++ src/application/controllers/index.ts
@@ -1,3 +1,2 @@
 export * from './list-pokemon.controller'
 export * from './find-pokemon-by-id.controller'
-export * from './find-pokemon-by-type.controller'
