+++ src/main/factories/controllers/list-pokemon.controller.factory.ts
@@ -1,6 +1,6 @@
-import { makeListPokemonService } from '@/main/factories/services'
+import { makeFindPokemonByTypeService, makeListPokemonService } from '@/main/factories/services'
 import { ListPokemonController } from '@/application/controllers'
 
 export const makeListPokemonController = (): ListPokemonController => {
-  return new ListPokemonController(makeListPokemonService())
+  return new ListPokemonController(makeListPokemonService(), makeFindPokemonByTypeService())
 }
