+++ src/main/factories/services/find-pokemon-by-type.service.factory.ts
@@ -1,0 +1,6 @@
+import { makePokemonRepository } from '@/main/factories/repositories'
+import { FindPokemonByTypeService } from '@/domain/services/pokemon'
+
+export const makeFindPokemonByTypeService = (): FindPokemonByTypeService => {
+  return new FindPokemonByTypeService(makePokemonRepository())
+}
