+++ src/main/factories/services/index.ts
@@ -1,2 +1,3 @@
 export * from './list-pokemon.service.factory'
 export * from './find-pokemon-by-id.service.factory'
+export * from './find-pokemon-by-type.service.factory'
